Add slippage buffer to cross exchange market making
Hi,

this is a request to add a slippage buffer parameter to the cross exchange market making strategy. I have received this code from Vik and tested it. Vik gave me permission to push this parameter trough so it can be used by all users.

This strategy limits the chance of an unfilled taker order by placing the taker order x% higher or lower than the min profit target level taker price. Setting this parameter to a high percentage will result in a "market order" type order.

Two important adjustments still need to be done to the code in order for this parameter to work properly.

- for maker order size calculation, take into account the slippage size. Only applicable for maker sell orders.
- add average maker fill price and taker fill price to terminals log for easy profit checking as hedging orders can now be executed unprofitably.