(feat) Refactor secure configs stroing
This PR refactors the way we store secure configs.

For strategies, the secure configs can be stored alongside
the rest of the strategy configs. They will be loaded with
the strategy configs and have the same lifetime.

For connectors, the login values / connector configs are
loaded on startup and are stored in the `Security` object
as previously. This allows for the export keys command
and several other parts of the code to continue working
as expected.

Connector configs are now stored in the same format as
strategy configs, and the secure values are no longer
separated in individual files.

All connectors have been refactored to this new approach to
configs storing.

The secure configs are no longer stored in the global
config map.