(feat) add Cross Exchange Mining Strategy
The cross exchange mining strategy sets orders on a maker exchange and seeks to profit off of the difference in the spread between taker and maker exchanges. The strategy is similar to the cross exchange market making strategy however it is more reliable in ensuring orders on the taker side are filled and assets remain hedged and balanced across exchanges. This is done by instead of looking for fill events on the maker side and attempting a trade on the taker (Which can fail) it constantly surveys the balance of currency across exchanges and buys or sells the currency on the most profitable exchange when a fill on the maker side occurs. This process is continuous and will not allow the hedge to fail.