(fix) stop hanging order cancellation endless loop
When an order is filled, then afterward is added as a hanging order, or when
an order is not created by the exchange and instead we receive a
MarketOrderFailureEvent, and in the meantime it is added as a hanging order,
then when the hanging order tracker detects that the order should be cancelled,
it goes into a cancellation loop, attempting to cancel the order every second,
forever.

The API call to cancel the order is not made repeatedly because hummingbot
knows that the order is not active.  However, a message is added to the logs
repeatedly, saying that the order is to be cancelled.

This patch updates the internal state of the hanging orders tracker so that
when an inactive order is set to be cancelled, instead of looping forever,
the order is removed from the list of hanging orders, as if it had been
cancelled.  This prevents the logs from filling up with repeated cancellation
lines.